move fetch groups, fetch group members and add group members queries
