+++ src/utils/types/friend.rs
@@ -1,0 +1,5 @@
+pub struct Friend {
+    pub id: i32,
+    pub username: String,
+    pub profile_picture: Option<String>,
+}
